remove WalletBasic
