+++ playground/nextjs-app-router/components/demo/Wallet.tsx
@@ -10,7 +10,6 @@ import { color } from '@coinbase/onchainkit/theme';
 import {
   ConnectWallet,
   Wallet,
-  WalletBasic,
   WalletDropdown,
   WalletDropdownBasename,
   WalletDropdownDisconnect,
@@ -25,31 +24,29 @@ function WalletComponent() {
   return (
     <div className="flex justify-end">
       <Wallet>
-        <WalletBasic>
-          <ConnectWallet text="Connect Wallet">
-            <Avatar address={address} className="h-6 w-6" />
+        <ConnectWallet text="Connect Wallet">
+          <Avatar address={address} className="h-6 w-6" />
+          <Name />
+        </ConnectWallet>
+        <WalletDropdown>
+          <Identity className="px-4 pt-3 pb-2">
+            <Avatar />
             <Name />
-          </ConnectWallet>
-          <WalletDropdown>
-            <Identity className="px-4 pt-3 pb-2">
-              <Avatar />
-              <Name />
-              <Address className={color.foregroundMuted} />
-              <EthBalance />
-              <Socials />
-            </Identity>
-            <WalletDropdownBasename />
-            <WalletDropdownLink
-              icon="wallet"
-              href="https://keys.coinbase.com"
-              target="_blank"
-            >
-              Wallet
-            </WalletDropdownLink>
-            <WalletDropdownFundLink />
-            <WalletDropdownDisconnect />
-          </WalletDropdown>
-        </WalletBasic>
+            <Address className={color.foregroundMuted} />
+            <EthBalance />
+            <Socials />
+          </Identity>
+          <WalletDropdownBasename />
+          <WalletDropdownLink
+            icon="wallet"
+            href="https://keys.coinbase.com"
+            target="_blank"
+          >
+            Wallet
+          </WalletDropdownLink>
+          <WalletDropdownFundLink />
+          <WalletDropdownDisconnect />
+        </WalletDropdown>
       </Wallet>
     </div>
   );
