+++ src/wallet/components/Wallet.test.tsx
@@ -3,7 +3,6 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { useOutsideClick } from '../../ui/react/internal/hooks/useOutsideClick';
 import { ConnectWallet } from './ConnectWallet';
 import { Wallet } from './Wallet';
-import { WalletBasic } from './WalletBasic';
 import { WalletDropdown } from './WalletDropdown';
 import { type WalletProviderReact, useWalletContext } from './WalletProvider';
 
@@ -53,12 +52,10 @@ describe('Wallet Component', () => {
 
     render(
       <Wallet>
-        <WalletBasic>
-          <ConnectWallet />
-          <WalletDropdown>
-            <div>Wallet Dropdown</div>
-          </WalletDropdown>
-        </WalletBasic>
+        <ConnectWallet />
+        <WalletDropdown>
+          <div>Wallet Dropdown</div>
+        </WalletDropdown>
       </Wallet>,
     );
 
@@ -81,12 +78,10 @@ describe('Wallet Component', () => {
 
     render(
       <Wallet>
-        <WalletBasic>
-          <ConnectWallet />
-          <WalletDropdown>
-            <div>Wallet Dropdown</div>
-          </WalletDropdown>
-        </WalletBasic>
+        <ConnectWallet />
+        <WalletDropdown>
+          <div>Wallet Dropdown</div>
+        </WalletDropdown>
       </Wallet>,
     );
 
@@ -111,12 +106,10 @@ describe('Wallet Component', () => {
 
     render(
       <Wallet>
-        <WalletBasic>
-          <ConnectWallet />
-          <WalletDropdown>
-            <div>Wallet Dropdown</div>
-          </WalletDropdown>
-        </WalletBasic>
+        <ConnectWallet />
+        <WalletDropdown>
+          <div>Wallet Dropdown</div>
+        </WalletDropdown>
       </Wallet>,
     );
