+++ src/wallet/components/Wallet.tsx
@@ -1,13 +1,29 @@
-import { useIsMounted } from '../../core-react/internal/hooks/useIsMounted';
-import { useTheme } from '../../core-react/internal/hooks/useTheme';
-import { cn } from '../../styles/theme';
-import { useOutsideClick } from '../../ui/react/internal/hooks/useOutsideClick';
+import { useIsMounted } from '@/core-react/internal/hooks/useIsMounted';
+import { useTheme } from '@/core-react/internal/hooks/useTheme';
+import { findComponent } from '@/core-react/internal/utils/findComponent';
+import { cn } from '@/styles/theme';
+import { useOutsideClick } from '@/ui-react/internal/hooks/useOutsideClick';
+import { Children, useMemo } from 'react';
 import type { WalletReact } from '../types';
 import { WalletProvider, useWalletContext } from './WalletProvider';
+import { ConnectWallet } from './ConnectWallet';
+import { WalletDropdown } from './WalletDropdown';
 
 function WalletContent({ children, className }: WalletReact) {
   const componentTheme = useTheme();
-  const { handleClose, containerRef: walletContainerRef } = useWalletContext();
+  const {
+    isOpen,
+    handleClose,
+    containerRef: walletContainerRef,
+  } = useWalletContext();
+
+  const { connect, dropdown } = useMemo(() => {
+    const childrenArray = Children.toArray(children);
+    return {
+      connect: childrenArray.find(findComponent(ConnectWallet)),
+      dropdown: childrenArray.find(findComponent(WalletDropdown)),
+    };
+  }, [children]);
 
   useOutsideClick(walletContainerRef, handleClose);
 
@@ -16,7 +32,10 @@ function WalletContent({ children, className }: WalletReact) {
       ref={walletContainerRef}
       className={cn('relative w-fit shrink-0', componentTheme, className)}
     >
-      {children}
+      <div className={cn('relative w-fit shrink-0', className)}>
+        {connect}
+        {isOpen && dropdown}
+      </div>
     </div>
   );
 }
