+++ src/wallet/components/WalletDefault.tsx
@@ -9,7 +9,6 @@ import {
 import { ConnectWallet } from './ConnectWallet';
 import { ConnectWalletText } from './ConnectWalletText';
 import { Wallet } from './Wallet';
-import { WalletBasic } from './WalletBasic';
 import { WalletDropdown } from './WalletDropdown';
 import { WalletDropdownDisconnect } from './WalletDropdownDisconnect';
 import { WalletDropdownLink } from './WalletDropdownLink';
@@ -17,29 +16,27 @@ import { WalletDropdownLink } from './WalletDropdownLink';
 export function WalletDefault() {
   return (
     <Wallet>
-      <WalletBasic>
-        <ConnectWallet>
-          <ConnectWalletText>Connect Wallet</ConnectWalletText>
-          <Avatar className="h-6 w-6" />
+      <ConnectWallet>
+        <ConnectWalletText>Connect Wallet</ConnectWalletText>
+        <Avatar className="h-6 w-6" />
+        <Name />
+      </ConnectWallet>
+      <WalletDropdown>
+        <Identity className="px-4 pt-3 pb-2">
+          <Avatar />
           <Name />
-        </ConnectWallet>
-        <WalletDropdown>
-          <Identity className="px-4 pt-3 pb-2">
-            <Avatar />
-            <Name />
-            <Address className={color.foregroundMuted} />
-            <EthBalance />
-          </Identity>
-          <WalletDropdownLink
-            icon="wallet"
-            href="https://keys.coinbase.com"
-            target="_blank"
-          >
-            Wallet
-          </WalletDropdownLink>
-          <WalletDropdownDisconnect />
-        </WalletDropdown>
-      </WalletBasic>
+          <Address className={color.foregroundMuted} />
+          <EthBalance />
+        </Identity>
+        <WalletDropdownLink
+          icon="wallet"
+          href="https://keys.coinbase.com"
+          target="_blank"
+        >
+          Wallet
+        </WalletDropdownLink>
+        <WalletDropdownDisconnect />
+      </WalletDropdown>
     </Wallet>
   );
 }
