+++ src/wallet/index.ts
@@ -2,7 +2,6 @@
 export { ConnectWallet } from './components/ConnectWallet';
 export { ConnectWalletText } from './components/ConnectWalletText';
 export { Wallet } from './components/Wallet';
-export { WalletBasic } from './components/WalletBasic';
 export { WalletDefault } from './components/WalletDefault';
 export { WalletDropdown } from './components/WalletDropdown';
 export { WalletDropdownBasename } from './components/WalletDropdownBasename';
